requirements.txt
python-dotenv
google-genai
yfinance>=0.2.18
pandas>=1.5.0
aiohttp>=3.8.0
asyncio
requests>=2.28.0
numpy>=1.21.0